feat: moved storage types to shared package
