+++ apps/api/src/app/storage/storage.controller.ts
@@ -1,5 +1,5 @@
 import { ClassSerializerInterceptor, Controller, Get, Query, UseGuards, UseInterceptors } from '@nestjs/common';
-import { IJwtPayload } from '@novu/shared';
+import { IJwtPayload, MimeTypesEnum } from '@novu/shared';
 import { GetSignedUrl } from './usecases/get-signed-url/get-signed-url.usecase';
 import { GetSignedUrlCommand } from './usecases/get-signed-url/get-signed-url.command';
 import { UserSession } from '../shared/framework/user.decorator';
@@ -48,7 +48,7 @@ export class StorageController {
   @ExternalApiAccessible()
   async profilePictureSignedUrl(
     @UserSession() user: IJwtPayload,
-    @Query('extension') extension: string
+    @Query('extension') extension: MimeTypesEnum
   ): Promise<UploadUrlResponse> {
     return await this.getUserProfilePictureSignedUrlUsecase.execute(
       GetUserProfilePictureSignedUrlCommand.create({
