+++ apps/api/src/app/storage/usecases/get-user-profile-picture-signed-url/get-user-profile-picture-signed-url.command.ts
@@ -1,19 +1,9 @@
-import { IsEnum, IsIn, IsString } from 'class-validator';
+import { MimeTypesEnum } from '@novu/shared';
+import { IsEnum, IsString } from 'class-validator';
 import { EnvironmentWithUserCommand } from '../../../shared/commands/project.command';
 
-export enum MimeTypesEnum {
-  JPEG = 'jpeg',
-  PNG = 'png',
-  JPG = 'jpg',
-}
 export class GetUserProfilePictureSignedUrlCommand extends EnvironmentWithUserCommand {
   @IsString()
   @IsEnum(MimeTypesEnum)
   extension: string;
 }
-
-export const MIME_TYPES_LOOKUP: Record<MimeTypesEnum, string> = {
-  [MimeTypesEnum.JPEG]: 'image/jpeg',
-  [MimeTypesEnum.PNG]: 'image/png',
-  [MimeTypesEnum.JPG]: 'image/jpeg',
-};
