+++ libs/shared/src/types/index.ts
@@ -23,3 +23,4 @@ export * from './auth';
 export * from './timezones';
 export * from './cron';
 export * from './product-features';
+export * from './storage';
